Fix dateSearch

+++ backend/Diabetia/Diabetia.Domain/Entities/Reporting/GlucoseMeasurement.cs
@@ -4,6 +4,6 @@ namespace Diabetia.Domain.Entities.Reporting
     public class GlucoseMeasurement
     {
         public DateTime MeasurementDate { get; set; }
-        public int GlucoseLevel { get; set; }
+        public decimal GlucoseLevel { get; set; }
     }
 }
